+++ .eslintrc.js
@@ -4,8 +4,7 @@ module.exports = {
     node: true
   },
   extends: [
-    'plugin:vue/recommended',
-    '@vue/standard'
+    'typescript'
   ],
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
@@ -14,8 +13,6 @@ module.exports = {
     'quotes': ['error', 'single', { 'allowTemplateLiterals': true }],
     'semi': ['error', 'always'],
     'space-before-function-paren': ['error', 'always'],
-    'vue/component-name-in-template-casing': ['error', 'PascalCase'],
-    'vue/eqeqeq': ['error', 'always', { 'null': 'ignore' }]
   },
   parserOptions: {
     parser: 'babel-eslint'