+++ test-package/import.test.ts
@@ -1,0 +1,14 @@
+// import { eq } from '../dist/index.js';
+
+
+// import { describe, it } from "mocha"
+// import { expect, assert } from "chai";
+
+// describe('package', () => {
+
+
+//   it('should build validation tree', () => {
+//     console.log(eq);
+//   });
+
+// });